+ blueprints, список транзакций для пользователя
